requirements.txt
numpy
pycuda
opencv-python
google-genai
matplotlib  
jupyter
streamlit
plotly
